pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-odds"
version = "0.1.0"
description = "Machine learning system for sports betting edge detection"
requires-python = ">=3.11"
dependencies = [
    "xgboost>=2.0.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.setuptools.packages.find]
include = ["ml_odds*"]
